pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Nikita <example@mail>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.1"
black = "^25.1.0"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
celery = "^5.5.2"
redis = "^6.2.0"
django-celery-beat = "^2.8.1"
drf-yasg = "^1.21.10"
flake8 = "^7.2.0"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
coverage = "^7.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
